+++ test/Raven.CodeAnalysis.Tests/Diagnostics/MemberCanBePrivateAnalyzerTests.cs
@@ -55,4 +55,5 @@ class Counter {
 
         verifier.Verify();
     }
+
 }